Added Nonce field to HTTP request for clicking buttons
Nonce field seemed unimportant, but not including it sometimes results in an error 400. After inclusion of this, the request payload exactly matches the payload when a button is pressed manually.
Nonce is a discord snowflake, which is a combination of time in seconds since discord epoch in binary combined with some other process IDs converted into an int64.
In case of our payload, the process ID's and other internal data is just a bunch of 0's.